refactor:上移PendingAtom相关数据模型
