+++ src/hivememory/engines/memory_compiler/compiler.py
@@ -86,9 +86,8 @@ class MemoryCompiler:
         target: MemoryCompileTarget,
         options: MemoryCompileOptions,
     ) -> CompiledMemoryArtifact:
-        from hivememory.alice.runtime.models import PendingAtom
         from hivememory.alice.runtime.resolver import ResolveResult
-        from hivememory.engines.generation.models import PendingAtomSettlement
+        from hivememory.core.models.pending import PendingAtom, PendingAtomSettlement
 
         if isinstance(source, MemoryAtom):
             return compile_memory_atom(source, target, options)
@@ -100,7 +99,7 @@ class MemoryCompiler:
             return compile_resolve_result(source, target, options)
 
         if isinstance(source, PendingAtomSettlement):
-            from hivememory.alice.runtime.pending_atom_state import PendingAtomResolution
+            from hivememory.core.models.pending import PendingAtomResolution
 
             kind = (
                 "discarded"
